# app/Http/Requests/Api/Product/ProductUpdateRequest.php
<?php

namespace App\Http\Requests\Api\Product;

use App\Http\Requests\Api\ApiValidationRequest;
use OpenApi\Attributes as OA;

#[OA\Schema(
    schema: 'ProductUpdateRequest',
    properties: [
        new OA\Property(property: 'name', type: 'string'),
        new OA\Property(property: 'description', type: 'string'),
        new OA\Property(property: 'price', type: 'number'),
        new OA\Property(property: 'stock_quantity', type: 'integer'),
    ],
    type: 'object'
)]
class ProductUpdateRequest extends ApiValidationRequest
{
    public function rules(): array
    {
        return [
            'name' => 'nullable|string|max:255',
            'description' => 'nullable|string|max:500',
            'price' => 'nullable|decimal:0,2|between:0,99999999.99',
            'stock_quantity' => 'nullable|integer',
        ];
    }
}